Drop a no-op cast in the unpacker's integer token reader
`mpack_unpack_sint` already answers `mpack_sintmax_t`, which is
`Integer`; the `as` was the transpile's and the cast deny does not
flag a same-type one. Removing it makes the two arms of
`unpack_integer_token` symmetric, which is what `rpcmutate`'s
`rpc-integer-token-sign` anchor swaps.

+++ crates/nvim/src/msgpack_rpc/unpacker/mod.rs
@@ -726,7 +726,7 @@ pub(super) fn unpack_integer_token(tok: mpack_token_t) -> Option<Integer> {
     if tok.type_0 == TOKEN_UINT {
         Some(mpack_unpack_uint(tok).cast_signed())
     } else if tok.type_0 == TOKEN_SINT {
-        Some(mpack_unpack_sint(tok) as Integer)
+        Some(mpack_unpack_sint(tok))
     } else {
         None
     }
